Add login capabilities
Not yet entirely sure which vulnerabilities to do, or what sort
of application to make (though that's not too important anyway).
Want to figure out how to use Django effectively too. However,
for the purposes of the assignment, some checks will probably
need to be disabled, or from the opposite angle, I will have to
explicitly use insecure methods in places, and then
use the standard Django variant for the secure variant. Naturally,
this means that the "fix" is quite simple in many cases.
Then again, CSRF is also allowed as a flaw, despite not being
in either of the lists and also being supposedly very simple to
deal with in Django.

Basically just wondering about how interesting the flaws/fixes need to
be. For example, broken access control is literally the difference
between "@login_required" and checking the correct user, and not doing that.