Fix brute force prevention
Realised that 1. it still tries to authenticate the username and password
in all cases, where I thought originally that it just ensures that
the format of the username and password is okay ("clean" doesn't exactly
mean testing authentication, and "form has no errors" does not mean authentication
succeeded either, so somewhat misleading naming. Although, clearly
I couldn't do "unsuccessful" login when just validating the form, so
I guess it is on me, and I must've realised that it would also
authenticate), and 2. letting the user to login when they give
the correct details obviously makes the point of the system moot, so a bit
of a moronic idea there. So, blocking attempts before any validation or
authentication is made is obviously the correct way to do it.